+++ DirectShowFilters/TsWriter/source/FileWriter.cpp
@@ -61,7 +61,7 @@ HRESULT FileWriter::SetFileName(LPCWSTR pszFileName)
   // Is this a valid filename supplied
   CheckPointer(pszFileName,E_POINTER);
 
-  long length = wcslen(pszFileName);
+  long length = (long)wcslen(pszFileName);
 
   if(length > MAX_PATH)
     return ERROR_FILENAME_EXCED_RANGE;
